// src/app/interfaces/ICocktail.ts
export interface IDrink {
    dateModified: Date;
    idDrink: string;
    strAlcoholic: string;
    strCategory: string;
    strCreativeCommonsConfirmed: string;
    strDrink: string;
    strDrinkAlternate: string;
    strDrinkDE: string;
    strDrinkES: string;
    strDrinkFR: string;
    strDrinkThumb: string;
    'strDrinkZH-HANS': string;
    'strDrinkZH-HANT': string;
    strGlass: string;
    strIBA: string;
    strImageAttribution: string;
    strImageSource: string;
    strIngredient1: string;
    strIngredient2: string;
    strIngredient3: string;
    strIngredient4: string;
    strIngredient5: string;
    strIngredient6: string;
    strIngredient7: string;
    strIngredient8: string;
    strIngredient9: string;
    strIngredient10: string;
    strIngredient11: string;
    strIngredient12: string;
    strIngredient13: string;
    strIngredient14: string;
    strIngredient15: string;
    strInstructions: string;
    strInstructionsDE: string;
    strInstructionsES: string;
    strInstructionsFR: string;
    'strInstructionsZH-HANS': string;
    'strInstructionsZH-HANT': string;
    strTags: string;
    strVideo: string;
    ingredients: string[];
    hasIngredient: boolean;
    hasCategory: boolean;
}